header cleanup: remove utils_fp.(cpp|h)
