+++ include/cufinufft/cudeconvolve.h
@@ -1,7 +1,6 @@
 #ifndef __CUDECONVOLVE_H__
 #define __CUDECONVOLVE_H__
 
-#include "cufinufft/contrib/utils_fp.h"
 #include <cufinufft_eitherprec.h>
 
 __global__ void Deconvolve_1d(int ms, int nf1, int fw_width, CUCPX *fw, CUCPX *fk, CUFINUFFT_FLT *fwkerhalf1);
