+++ test/cufinufft3d2_test.cu
@@ -1,14 +1,13 @@
-#include "cufinufft/profile.h"
 #include <complex>
 #include <helper_cuda.h>
 #include <iomanip>
 #include <iostream>
-#include <math.h>
+#include <random>
 
 #include <cufinufft_eitherprec.h>
+#include <cufinufft/utils.h>
+#include <cufinufft/profile.h>
 
-#include "cufinufft/contrib/utils.h"
-#include "cufinufft/contrib/utils_fp.h"
 
 int main(int argc, char *argv[]) {
     int N1, N2, N3, M;
@@ -64,6 +63,10 @@ int main(int argc, char *argv[]) {
     checkCudaErrors(cudaMalloc(&d_c, M * sizeof(CUCPX)));
     checkCudaErrors(cudaMalloc(&d_fk, N1 * N2 * N3 * sizeof(CUCPX)));
 
+    std::default_random_engine eng(1);
+    std::uniform_real_distribution<CUFINUFFT_FLT> dist11(-1, 1);
+    auto randm11 = [&eng, &dist11]() { return dist11(eng); };
+    
     // Making data
     for (int i = 0; i < M; i++) {
         x[i] = M_PI * randm11(); // x in [-pi,pi)
